refactor(core): unify metafile rasterizers on the shared createAuxCanvas fallback
renderWmfToBitmap allocated `new OffscreenCanvas` unconditionally, so a WMF
blip hard-required OffscreenCanvas and threw where it was absent (main-thread
runtimes without it). renderEmfToBitmap and blitDibToCtx each guarded with
their own `typeof OffscreenCanvas === 'undefined'` early-out and then
allocated directly — three variants of the same "prefer OffscreenCanvas, else
detached <canvas>, else give up" logic that shape/effects.ts already
centralized in createAuxCanvas.

Route all three through createAuxCanvas. To avoid an unnatural image → shape
dependency, move createAuxCanvas (and its AuxCanvas/AuxContext types) to a new
leaf module core/canvas/aux-canvas.ts and re-export it from shape/effects.ts,
so the public API (`createAuxCanvas` via effects and the core barrel) is
unchanged. wmf/emf's `createImageBitmap(canvas)` and dib's `drawImage(tmp,…)`
both accept an HTMLCanvasElement, so the <canvas> fallback is type-safe; a
null result maps to each function's existing "missing image" / skip path.

This also makes EMF and DIB blits work on the main thread (detached <canvas>)
where before they silently returned null/false without OffscreenCanvas — a
strict superset of the old behavior. The worker-only
`new OffscreenCanvas + transferToImageBitmap` paths in docx/pptx/xlsx are NOT
touched: those legitimately require OffscreenCanvas (bitmap transfer to the
main thread). node's installOffscreenCanvasShim probe stays green.
